+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Set the proper origin remote on working directory repositories.
+
 # 0.23.0
 
 * Always fetch from the remote before a task to ensure any newly pushed tag is present.